src_original/network_monitoring.py
# Project name: Secure7NetGuard_1.01_Offline ;
# Author: Nelsomar Barros ;
# Github: https://github.com/SecureLogic7/Secure7NetGuard ;
# Contact: nelsom.one8@gmail.com .

import pcapplusplus as ppp

class NetworkMonitoring:
    def __init__(self):
        self.packets = []
        self._offline = False

    def functionality(self):
        return "expected_result"

    def monitor_network(self):
        pcap = ppp.PcapLiveDeviceList.getInstance().getPcapLiveDeviceByName("eth0")
        pcap.open()
        pcap.startCapture()

        try:
            while True:
                packet = pcap.getNextPacket()

                if packet:
                    self.analyze_packet(packet)
                    self.packets.append(packet)

        except KeyboardInterrupt:
            pcap.stopCapture()
            pcap.close()

    def analyze_packet(self, packet):
        print("Packet analyzed:", packet)

    def get_network_status(self):
        if self._offline:
            return "offline"
        else:
            return "online"

    def set_offline_mode(self, offline: bool):
        self._offline = offline

    def get_ip_address(self):
        return "192.168.1.1"

    def get_network_speed(self):
        return 100.0
