requirements.txt
Flask==3.1.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
SQLAlchemy==1.4.23
Werkzeug==3.1.3
cryptography
pytest
requests
stripe
Pillow
python-dotenv
